+++ example/lib/src/capture_frame_sample.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 class CaptureFrameSample extends StatefulWidget {
   @override
