+++ example/lib/src/device_enumeration_sample.dart
@@ -1,10 +1,10 @@
 import 'dart:core';
-import 'package:collection/collection.dart';
 
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:permission_handler/permission_handler.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 class VideoSize {
   VideoSize(this.width, this.height);
