+++ example/lib/src/get_display_media_sample.dart
@@ -3,8 +3,8 @@ import 'dart:core';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_background/flutter_background.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:flutter_webrtc_example/src/widgets/screen_select_dialog.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 /*
  * getDisplayMedia sample
