+++ example/lib/src/get_user_media_sample.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 /*
  * getUserMedia sample
