+++ example/lib/src/get_user_media_sample_web.dart
@@ -4,7 +4,7 @@ import 'dart:html' as html;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 /*
  * getUserMedia sample
