+++ example/lib/src/loopback_sample_with_get_stats.dart
@@ -1,7 +1,7 @@
 import 'dart:core';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 class LoopBackSampleWithGetStats extends StatefulWidget {
   static String tag = 'loopback_sample_with_get_stats';
