+++ example/lib/src/widgets/screen_select_dialog.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart';
 
 class ThumbnailWidget extends StatefulWidget {
   const ThumbnailWidget(
