+++ example/linux/flutter/generated_plugin_registrant.cc
@@ -6,10 +6,10 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <flutter_webrtc/flutter_web_r_t_c_plugin.h>
+#include <stream_webrtc_flutter/flutter_web_r_t_c_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
-  g_autoptr(FlPluginRegistrar) flutter_webrtc_registrar =
+  g_autoptr(FlPluginRegistrar) stream_webrtc_flutter_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterWebRTCPlugin");
-  flutter_web_r_t_c_plugin_register_with_registrar(flutter_webrtc_registrar);
+  flutter_web_r_t_c_plugin_register_with_registrar(stream_webrtc_flutter_registrar);
 }
