+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,10 +5,10 @@
 import FlutterMacOS
 import Foundation
 
-import flutter_webrtc
 import path_provider_foundation
+import stream_webrtc_flutter
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  FlutterWebRTCPlugin.register(with: registry.registrar(forPlugin: "FlutterWebRTCPlugin"))
   PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
+  FlutterWebRTCPlugin.register(with: registry.registrar(forPlugin: "FlutterWebRTCPlugin"))
 }
