+++ example/pubspec.yaml
@@ -12,12 +12,12 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_background: ^1.0.0
-  flutter_webrtc:
-    path: ../
   # Required for MediaRecorder example
   path_provider: ^2.0.2
   permission_handler: ^11.3.1
   sdp_transform: ^0.3.2
+  stream_webrtc_flutter:
+    path: ../
 
 
 dev_dependencies:
