+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,12 +6,12 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <flutter_webrtc/flutter_web_r_t_c_plugin.h>
 #include <permission_handler_windows/permission_handler_windows_plugin.h>
+#include <stream_webrtc_flutter/flutter_web_r_t_c_plugin.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  FlutterWebRTCPluginRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("FlutterWebRTCPlugin"));
   PermissionHandlerWindowsPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("PermissionHandlerWindowsPlugin"));
+  FlutterWebRTCPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FlutterWebRTCPlugin"));
 }
