+++ example/windows/flutter/generated_plugins.cmake
@@ -3,8 +3,8 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  flutter_webrtc
   permission_handler_windows
+  stream_webrtc_flutter
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
