+++ ios/stream_webrtc_flutter.podspec
@@ -2,15 +2,15 @@
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html
 #
 Pod::Spec.new do |s|
-  s.name             = 'flutter_webrtc'
-  s.version          = '0.12.2'
+  s.name             = 'stream_webrtc_flutter'
+  s.version          = '0.12.2+1'
   s.summary          = 'Flutter WebRTC plugin for iOS.'
   s.description      = <<-DESC
 A new flutter plugin project.
                        DESC
-  s.homepage         = 'https://github.com/cloudwebrtc/flutter-webrtc'
+  s.homepage         = 'https://github.com/GetStream/webrtc-flutter'
   s.license          = { :file => '../LICENSE' }
-  s.author           = { 'CloudWebRTC' => 'duanweiwei1982@gmail.com' }
+  s.author           = { 'getstream.io' => 'support@getstream.io' }
   s.source           = { :path => '.' }
   s.source_files = 'Classes/**/*'
   s.public_header_files = 'Classes/**/*.h'
