+++ macos/flutter_webrtc.podspec
@@ -3,14 +3,14 @@
 #
 Pod::Spec.new do |s|
   s.name             = 'flutter_webrtc'
-  s.version          = '0.12.2'
+  s.version          = '0.12.2+1'
   s.summary          = 'Flutter WebRTC plugin for macOS.'
   s.description      = <<-DESC
 A new flutter plugin project.
                        DESC
-  s.homepage         = 'https://github.com/cloudwebrtc/flutter-webrtc'
+  s.homepage         = 'https://github.com/GetStream/webrtc-flutter'
   s.license          = { :file => '../LICENSE' }
-  s.author           = { 'CloudWebRTC' => 'duanweiwei1982@gmail.com' }
+  s.author           = { 'getstream.io' => 'support@getstream.io' }
   s.source           = { :path => '.' }
   s.source_files     = ['Classes/**/*']
